requirements.txt
aiohttp
configparser
future
fuzzywuzzy
psutil
pytz

# Additional Packages for export Functionality
pandas>=2.0.0

# Extra libraries are installed when needed using pipmaster

pipmaster
pydantic
python-dotenv

# Unicode Collation Algorithm for proper Chinese sorting
pyuca

setuptools
tenacity

# LLM packages
tiktoken
xlsxwriter>=3.1.0
